requirements.txt
requests
pandas
lxml
beautifulsoup4
line-bot-sdk>=3.0.0
